feat: web wasm embeddings — int32 model surface + vendored ORT runtime
The light client gets real MiniLM instead of the FNV surrogate:
- model re-exported (tool/trim_model.py) with an INT32 input surface
  casting to int64 inside the graph: dart2js has no Int64List and the
  web plugin cannot up-convert; one artifact serves every platform.
  Quality gates and the int8 'warm up' beacon are UNCHANGED (python:
  EN 100%/0.5%, RU 93%/2.9% at 0.45). Asset bumped to _v2 per the
  by-name cache rule.
- OnnxEmbeddingService: Int32List tensors everywhere; on web the model
  loads via rootBundle -> Blob URL (independent of asset URL layout),
  through a conditional-import helper (model_bytes_url_*).
- web/ort/: vendored onnxruntime-web 1.23.0 (ort.min.js + wasm) wired
  in index.html before the Flutter bootstrap — local-first, no CDN at
  runtime.
- main.dart: single ONNX-with-FNV-fallback path for native AND web.

Verified: analyze clean, 40/40 VM tests, python int8 gates green.
PENDING: test/onnx_web_test.dart (chrome-platform beacon parity) hung
in the headless harness on first run — investigation continues next;
native device beacon re-check also outstanding for the v2 asset.

+++ tool/gate_int8.py
@@ -19,7 +19,7 @@ session = ort.InferenceSession(
 
 
 def embed(text):
-    ids = np.array([tok.encode(text).ids], dtype=np.int64)
+    ids = np.array([tok.encode(text).ids], dtype=np.int32)
     hidden = session.run(
         ["last_hidden_state"],
         {"input_ids": ids, "attention_mask": np.ones_like(ids)},
